experiment: wgpu sync
