+++ ext/webgpu/02_idl_types.js
@@ -1,4 +1,4 @@
-// Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
+// Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 
 // @ts-check
 /// <reference path="../web/internal.d.ts" />
@@ -111,28 +111,35 @@
   webidl.converters["GPUFeatureName"] = webidl.createEnumConverter(
     "GPUFeatureName",
     [
-      "depth-clamping",
+      "depth-clip-control",
       "depth24unorm-stencil8",
       "depth32float-stencil8",
       "pipeline-statistics-query",
       "texture-compression-bc",
+      "texture-compression-etc2",
+      "texture-compression-astc",
       "timestamp-query",
+      "indirect-first-instance",
       // extended from spec
       "mappable-primary-buffers",
-      "sampled-texture-binding-array",
-      "sampled-texture-array-dynamic-indexing",
-      "sampled-texture-array-non-uniform-indexing",
+      "texture-binding-array",
+      "buffer-binding-array",
+      "storage-resource-binding-array",
+      "sampled-texture-and-storage-buffer-array-non-uniform-indexing",
+      "uniform-buffer-and-storage-buffer-texture-non-uniform-indexing",
       "unsized-binding-array",
       "multi-draw-indirect",
       "multi-draw-indirect-count",
       "push-constants",
       "address-mode-clamp-to-border",
-      "non-fill-polygon-mode",
-      "texture-compression-etc2",
-      "texture-compression-astc-ldr",
       "texture-adapter-specific-format-features",
       "shader-float64",
       "vertex-attribute-64bit",
+      "conservative-rasterization",
+      "vertex-writable-storage",
+      "clear-commands",
+      "spirv-shader-passthrough",
+      "shader-primitive-index",
     ],
   );
 
@@ -348,6 +355,44 @@
       "bc6h-rgb-float",
       "bc7-rgba-unorm",
       "bc7-rgba-unorm-srgb",
+      "etc2-rgb8unorm",
+      "etc2-rgb8unorm-srgb",
+      "etc2-rgb8a1unorm",
+      "etc2-rgb8a1unorm-srgb",
+      "etc2-rgba8unorm",
+      "etc2-rgba8unorm-srgb",
+      "eac-r11unorm",
+      "eac-r11snorm",
+      "eac-rg11unorm",
+      "eac-rg11snorm",
+      "astc-4x4-unorm",
+      "astc-4x4-unorm-srgb",
+      "astc-5x4-unorm",
+      "astc-5x4-unorm-srgb",
+      "astc-5x5-unorm",
+      "astc-5x5-unorm-srgb",
+      "astc-6x5-unorm",
+      "astc-6x5-unorm-srgb",
+      "astc-6x6-unorm",
+      "astc-6x6-unorm-srgb",
+      "astc-8x5-unorm",
+      "astc-8x5-unorm-srgb",
+      "astc-8x6-unorm",
+      "astc-8x6-unorm-srgb",
+      "astc-8x8-unorm",
+      "astc-8x8-unorm-srgb",
+      "astc-10x5-unorm",
+      "astc-10x5-unorm-srgb",
+      "astc-10x6-unorm",
+      "astc-10x6-unorm-srgb",
+      "astc-10x8-unorm",
+      "astc-10x8-unorm-srgb",
+      "astc-10x10-unorm",
+      "astc-10x10-unorm-srgb",
+      "astc-12x10-unorm",
+      "astc-12x10-unorm-srgb",
+      "astc-12x12-unorm",
+      "astc-12x12-unorm-srgb",
       "depth24unorm-stencil8",
       "depth32float-stencil8",
     ],
@@ -1131,7 +1176,7 @@
       defaultValue: "none",
     },
     {
-      key: "clampDepth",
+      key: "unclippedDepth",
       converter: webidl.converters["boolean"],
       defaultValue: false,
     },
@@ -1458,13 +1503,7 @@
   );
 
   // DICTIONARY: GPUCommandEncoderDescriptor
-  const dictMembersGPUCommandEncoderDescriptor = [
-    {
-      key: "measureExecutionTime",
-      converter: webidl.converters["boolean"],
-      defaultValue: false,
-    },
-  ];
+  const dictMembersGPUCommandEncoderDescriptor = [];
   webidl.converters["GPUCommandEncoderDescriptor"] = webidl
     .createDictionaryConverter(
       "GPUCommandEncoderDescriptor",
