+++ ext/webgpu/Cargo.toml
@@ -1,4 +1,4 @@
-# Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
+# Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 
 [package]
 name = "deno_webgpu"
@@ -7,15 +7,12 @@ authors = ["the Deno authors"]
 edition = "2021"
 license = "MIT"
 readme = "README.md"
-repository = "https://github.com/denoland/deno"
+repository = "https://github.com/gfx-rs/wgpu"
 description = "WebGPU implementation for Deno"
-
-[lib]
-path = "lib.rs"
 
 [dependencies]
 deno_core = { version = "0.114.0", path = "../../core" }
-serde = { version = "1.0.129", features = ["derive"] }
-tokio = { version = "1.10.1", features = ["full"] }
-wgpu-core = { version = "0.10.1", features = ["trace"] }
-wgpu-types = "0.10.0"
+serde = { version = "1.0", features = ["derive"] }
+tokio = { version = "1.10", features = ["full"] }
+wgpu-core = { version = "0.12.0", features = ["trace", "replay", "serde"] }
+wgpu-types = { version = "0.12.0", features = ["trace", "replay", "serde"] }
